app/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
py-healthcheck==1.10.1
pyparsing==2.4.7
pytest==6.2.4
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
